Revert "[Feat] #324 전체 상품 리스트"

See merge request s10-bigdata-recom-sub2/S10P22E106!109

Former-commit-id: 6f1e099f0c8614fd419d3f622ebbc213415b27a1

+++ back/src/main/java/com/msm/back/product/dto/ProductListRequestDto.java
@@ -16,5 +16,5 @@ public class ProductListRequestDto {
 
 	private Long lastId;
 
-	private int pageSize;
+	private Long pageSize;
 }
